+++ app/src/main/java/camp/nextstep/MyWebApplicationInitializer.java
@@ -1,6 +1,6 @@
 package camp.nextstep;
 
-import com.interface21.webmvc.servlet.mvc.tobe.DispatcherServlet;
+import com.interface21.webmvc.servlet.mvc.DispatcherServlet;
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
