+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/JsonResolver.java
@@ -1,6 +1,6 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.mvc;
 
-import com.interface21.webmvc.servlet.view.JsonView;
+import com.interface21.webmvc.servlet.mvc.view.JsonView;
 
 public class JsonResolver {
 	public static JsonView resolveJsonView() {
