+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/ViewResolver.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.mvc;
 
 public interface ViewResolver {
 	View resolveViewName(String viewName);
