+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/support/PathPatternUtilTest.java
@@ -1,5 +1,6 @@
-package com.interface21.webmvc.servlet.mvc.tobe.support;
+package com.interface21.webmvc.servlet.mvc.support;
 
+import com.interface21.webmvc.servlet.mvc.support.PathPatternUtil;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
